### 2023-06-11 #5534 @editorialbot
[PRE REVIEW]: Pycapacity: a real-time task-space capacity calculation package for robotics and biomechanics
**Submitting author:** @askuric (<a href="http://orcid.org/0000-0002-3323-4482">Antun Skuric</a>)
**Repository:** https://github.com/auctus-team/pycapacity
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.1
**Managing EiC:** Kevin M Moerman
