### 2023-06-12 #5534 @askuric
Hi @arfon,

Thanks for the heads-up. I've updated the paper with the proper ORCID.

The list of possible reviewers from the list that seem to correspond the most to the robotics and applied mathematics search
- [@sea-bass](https://reviewers.joss.theoj.org/reviewers/1543)
- [@jcarpent](https://reviewers.joss.theoj.org/reviewers/682)
- [@gfadini](https://reviewers.joss.theoj.org/reviewers/2189)
