### 2023-06-26 b47c34a
Fix affiliation index for the JOSS draft paper
Fix affiliation index for the JOSS draft paper
