### 2023-06-26 #5534 @Kevin-Mattheus-Moerman
@askuric the paper still has issues. I think perhaps this PR might fix it: https://github.com/auctus-team/pycapacity/pull/3 (`index: 1` was missing). Note you can update the paper here, and trouble shoot issues yourself using:  `@editorialbot generate pdf`
