+++ paper/paper.md
@@ -15,7 +15,8 @@ authors:
   - name: David Daney
     affiliation: 1
 affiliations:
- - name: INRIA, Bordeaux, France
+  - name: INRIA, Bordeaux, France
+    index: 1
 date: 23 May 2023
 bibliography: paper.bib
 ---
