### 2023-07-20 #5534 @gfadini
Hi, I am interested in this paper, but I am a bit busy lately and I will be
able to focus on it just starting from August.
How much time is usually left to submit the review?
Best,
Gabriele

Il giorno mer 19 lug 2023 alle ore 12:03 Dana Solav <
***@***.***> ha scritto:

> Hello @sea-bass <https://github.com/sea-bass>, @jcarpent
> <https://github.com/jcarpent>, @gfadini <https://github.com/gfadini>
> Would you be interested in reviewing this paper?
>
