### 2023-07-20 #5670 @editorialbot
[REVIEW]: Pycapacity: a real-time task-space capacity calculation package for robotics and biomechanics
**Submitting author:** @askuric (<a href="http://orcid.org/0000-0002-3323-4482">Antun Skuric</a>)
**Repository:** https://github.com/auctus-team/pycapacity
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.2
