### 2023-07-23 auctus-team/pycapacity#4 @sea-bass
[JOSS Review] Add "interesting" doc examples to the examples folder
The docs have lots of great examples of PyCapacity integrating with Peter Corke's robotics toolbox, Pinocchio, etc. However, these standalone Python files are not available in the repo itself -- you have to copy-paste them from the docs. The `examples` folder only has simple randomized robot examples.

Please consider adding these doc examples in as "official" examples so it's easy to run them directly without having to make a new file and copy the docs contents. You can also have the docs point to the path of these example files, for example `python3 examples/robotics_toolbox_example.py`

Totally understandable if users have to `pip install` a few extra dependencies, simply mark this in the comments for each example as appropriate. You could also share those anaconda / pip configs in the examples folder for the users to easily set up.

https://github.com/openjournals/joss-reviews/issues/5670
