### 2023-07-23 auctus-team/pycapacity#5 @sea-bass
[JOSS Review] Issues with Doc Examples
In addition to https://github.com/auctus-team/pycapacity/issues/4, I have some feedback on the examples themselves:

## Robotics Toolbox Example
I can't get the Swift example working in my Python 3.10 environment -- I keep getting this error when visualizing

```
Traceback (most recent call last):
  File "/home/sebastian/Workspace/pycapacity/.virtualenvs/pycapacity/lib/python3.10/site-packages/websockets/legacy/server.py", line 240, in handler
    await self.ws_handler(self)
  File "/home/sebastian/Workspace/pycapacity/.virtualenvs/pycapacity/lib/python3.10/site-packages/websockets/legacy/server.py", line 1186, in _ws_handler
    return await cast(
  File "/home/sebastian/Workspace/pycapacity/.virtualenvs/pycapacity/lib/python3.10/site-packages/swift/SwiftRoute.py", line 320, in serve
    await self.expect_message(websocket, expected)
  File "/home/sebastian/Workspace/pycapacity/.virtualenvs/pycapacity/lib/python3.10/site-packages/swift/SwiftRoute.py", line 325, in expect_message
    recieved = await websocket.recv()
  File "/home/sebastian/Workspace/pycapacity/.virtualenvs/pycapacity/lib/python3.10/site-packages/websockets/legacy/protocol.py", line 568, in recv
    await self.ensure_open()
  File "/home/sebastian/Workspace/pycapacity/.virtualenvs/pycapacity/lib/python3.10/site-packages/websockets/legacy/protocol.py", line 935, in ensure_open
    raise self.connection_closed_exc()
websockets.exceptions.ConnectionClosedOK: received 1001 (going away); then sent 1001 (going away)
```

Also, I had to additionally `pip install networkx` because it was another missing dependency not listed anywhere.

---

## Pinocchio Example
The `pip` instructions just say `pip install pin`, but actually you need other things like `example-robot-data`, `meshcat`, etc. Make sure this is consistent.

Also, the first (static) Meshcat example works for me but it's all blank... any ideas?
![image](https://github.com/auctus-team/pycapacity/assets/4603398/2fe2b833-1b70-4f3e-90bd-ed34a018a33d)

The animation one works fine...
![image](https://github.com/auctus-team/pycapacity/assets/4603398/ec146d13-ac25-4d18-a65e-1e23bdbc150c)

Actually, if in the first example I stick these lines at the end, the robot ends up rendering fine.

```python
while True:
    viz.display(q0)
```

---

## Four Link Planar Robot

If running from the top of the module, this import fails:
```
from four_link_utils import *
```

The doc says to copy files to a `four_link_utils.py`, but maybe worth having an easier way to just use the one that already exists in the examples folder? Or tell people to just define the necessary utils in the same example file?

https://github.com/openjournals/joss-reviews/issues/5670
