### 2023-07-23 auctus-team/pycapacity#6 @sea-bass
[JOSS Review] Paper Feedback
Thank you for submitting this cool toolbox! Here is my feedback on the paper.

## General Feedback
* I may debate the claim that a Python package should be used for real-time applications. If that was critical, then this package should have been implemented in C++, Rust, or something like that. I think Python is good for ease of use and prototyping, but I may pull back on some of these claims.
* In addition to the citations, which you already have, I think it would be helpful to have a brief introduction to these evaluation metrics (ellipsoids and polytopes) and what they do. As such, I think your "Physical Capacity Metrics" would be better served BEFORE the "Implemented Polytope Evaluation Algorithms" section. First say what these things are, and what they are used for, and then get into which algorithms are in PyCapacity.
* The tables of evaluation times need to have some notion of variance. Instead of saying things like `~5ms`, I (and readers) would be mostly interested in something like `5ms +/- 1ms`, as well as a brief description of how this data was collected (how many samples, etc.) -- see e.g., https://jhavl.github.io/mmc/ for an example.
* Please add a Conclusion section at the end
* Capitalize Python, Jacobian, and Cartesian throughout

---

## Summary
* Typo in line 7: "the robot's are required" --> "robots are often required"
* Typo in line 9: "One of ways to quantify" --> "One way to quantify"
* Line 11: Consider gender-neutral language in line 11 "where he lacks" --> "where they lack"
* Typo in line 13: "as todays collaborative" --> "as today's collaborative"
* Missing : in line 19 -- "Most of these metrics can be represented by two families of geometric shapes: ellipsoids and polytopes"
* Finish your summary by saying what this paper does: It introduces a new package named PyCapacity!

---

## Statement of Need
* More of a question: Have there been other tools available, in Python or otherwise, that do these ellipsoid/polytope calculations and visualizations? If so, you should list them and cite them. For example:
  * https://jhavl.github.io/mmc/
  * https://github.com/tecnalia-medical-robotics/manipulability_metrics
  * https://gitlab.com/KeerthiSagarSN/rospygradientpolytope
* Typo in line 27: "human musculoskeletal modes" -> "human musculoskeletal models"
* I really like that you demonstrated which existing publications already use your work. Very important!

---

## Physical Capacity Metrics
* Typo in line 83: "Reachable space approximation of in the desired horizon" -- of what? I think it is either missing a word, or the "of" should be deleted
* The caption in Figure 2 is missing a period after "visualized with `biorbd`".

https://github.com/openjournals/joss-reviews/issues/5670
