### 2023-07-23 auctus-team/pycapacity#7 @sea-bass
[JOSS Review] Add more descriptive function names to unit tests
Right now, all the unit test functions are named `test1()`, `test2()`, etc. which is not very descriptive.

Consider more readable names that communicate the purpose of the test, for example `test_invalid_polytope_x_min()`.

https://github.com/openjournals/joss-reviews/issues/5670
