### 2023-07-23 4d6420d
[JOSS REVIEW] Fix typos
