+++ docs/source/examples/randomised_human.md
@@ -1,7 +1,7 @@
 Randomized Human musculoskeletal models
-==================
+=======================================
 
-A simple example program for force polytope  evaluation of a randomised human musculoskeletal model. 
+A simple example program for force polytope evaluation of a randomised human musculoskeletal model. 
 Simply change the number of dof, number of forces and force limits and see how the calculation time and shape evaluates.
 
 ```python
@@ -30,10 +30,10 @@ fig = plt.figure(4)
 
 # draw faces and vertices
 plot_polytope(plot=plt, 
-            polytope=f_poly, 
-            label='force', 
-            edge_color='black', 
-            alpha = 0.4)
+              polytope=f_poly, 
+              label='force', 
+              edge_color='black', 
+              alpha = 0.4)
 
 plt.legend()
 plt.show()
