### 2023-07-23 #5670 @sea-bass
I've gone through this submission and it looks like a great set of tools! I think it largely meets all the necessary criteria close to being a state fit for submission, but I have a few things pending.

* I have some feedback on the paper itself. Most of it is minor, but I also have some bigger comments on the claims of "real-time" for a Python tool, and also requesting variance/error in the computation time metrics: https://github.com/auctus-team/pycapacity/issues/6
* The "interesting" examples (e.g., Robotics Toolbox, Pinocchio, etc.) are available in the docs markdown, but should also be ready to run in the repo itself: https://github.com/auctus-team/pycapacity/issues/4
* Some of the examples I tried had issues running. Maybe it's user error, or maybe something else, but please look at these: https://github.com/auctus-team/pycapacity/issues/5
* The unit tests look quite complete, but the functions all have names like `test1`, `test2`, etc. which could be made more descriptive to help users understand and contribute: https://github.com/auctus-team/pycapacity/issues/7
* I also found a lot of small typos in the tool as I was going through, so I've put up a PR: https://github.com/auctus-team/pycapacity/pull/8

Looking forward to collaborating on addressing these issues!
