+++ tests/testing_algos.py
@@ -2,7 +2,8 @@ import pycapacity.algorithms as algos
 import numpy as np
 
 
-def test1():
+# function testing if algotrithms are execute for random data
+def test_algos_random_data():
     L = 10
     n = 5
     m = 3
@@ -27,7 +28,7 @@ def test1():
 
 # write unit test for hyper_plane_shift_method
 # with random data and return true id no exception is raised
-def test2():
+def test_hpsm_random():
     L = 10
     n = 5
     m = 3
@@ -43,7 +44,7 @@ def test2():
 
 # write unit test for vertex_enumeration_auctus
 # with random data and return true id no exception is raised
-def test3():
+def test_vea_random():
     L = 10
     n = 5
     m = 3
@@ -59,7 +60,7 @@ def test3():
 
 #  write unit test for iterative_convex_hull_method
 # with random data and return true id no exception is raised
-def test4():
+def test_ichm_random():
     L = 10
     n = 5
     m = 3
@@ -74,9 +75,9 @@ def test4():
 
     assert True
 
-# unit test for iterative_convex_hull_method return true if expection is 
+# unit test for iterative_convex_hull_method return true if exception is 
 # raised when x_min and x_max are not the same size
-def test5():
+def test_ichm_random_xmin_xmax_not_same_size():
     L = 10
     n = 5
     m = 3
@@ -94,7 +95,7 @@ def test5():
 
 # unit test for iterative_convex_hull_method check if A nad B matrices not compatible
 # return true if expection is raised
-def test6():
+def test_ichm_random_A_B_not_compatible():
     L = 10
     n = 5
     m = 3
@@ -112,7 +113,7 @@ def test6():
 
 # unit test for iterative_convex_hull_method with singular matrix A
 # return true if expection is raised
-def test7():
+def test_ichm_random_A_singular():
     L = 10
     n = 5
     m = 3
@@ -133,7 +134,7 @@ def test7():
 
 # unit test for iterative_convex_hull_method no matrix A is identity matrix
 # return true if all is ok
-def test8():
+def test_ichm_random_A_identity():
     L = 10
     n = 5
     m = 3
@@ -150,7 +151,7 @@ def test8():
 
 # unit test for iterative_convex_hull_method with matrix B is identity matrix
 # return true if all is ok
-def test9():
+def test_ichm_random_B_identity():
     L = 10
     n = 5
     m = 3
@@ -167,7 +168,7 @@ def test9():
 
 # unit test for iterative_convex_hull_method with equality constraints G_eq and h_eq
 # return true if all is ok
-def test10():
+def test_ichm_random_G_eq_h_eq():
     np.random.seed(102)
     L = 10
     n = 5
@@ -187,7 +188,7 @@ def test10():
 
 # unit test iterative_convex_hull_method if matrix A has less rows than columns
 # return is exception is raised
-def test11():
+def test_ichm_random_A_less_rows_than_columns():
     L = 10
     n = 5
     m = 3
@@ -205,7 +206,7 @@ def test11():
     
 # unit test iterative_convex_hull_method if lower limits higher than upper limits
 # return is exception is raised
-def test12():
+def test_ichm_random_y_min_higher_than_y_max():
     L = 10
     n = 5
     m = 3
@@ -223,7 +224,7 @@ def test12():
 
 # unit test iterative_convex_hull_method with bias  
 # return true if all is ok
-def test13():
+def test_ichm_random_bias():
     L = 10
     n = 5
     m = 3
[…]
