### 2023-07-24 auctus-team/pycapacity#7 @askuric
Thanks for the feedback, I've updated the function names. I was trying to be as descriptive as possible in the comments, but I agree that the redundancy improves the clarity. :D
Thanks!
