### 2023-07-24 685bd61
added examples module with the FourLinkRobot + updated docs and exmaples for this + some more typos found
