+++ examples/4dof_robot_examples/acceleration_capacity.py
@@ -1,57 +1,58 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-import pycapacity.robot as capacity # robot capacity module
-import pycapacity.visual as visual # visualistion tools
+from pycapacity.robot import * # robot capacity module
+from pycapacity.visual import * # visualistion tools
 
-from four_link_utils import *
+# four link robot import
+from pycapacity.examples import FourLinkRobot
+
+# create the robot
+robot = FourLinkRobot()
 
 #joint positions q
-q  = np.random.rand(4)*np.pi/2
+q  = np.random.rand(4)*np.pi/3*2-1
 # joint torque limits tau
-tau_min = -np.ones((4,1))
-tau_max = np.ones((4,1))
-
+dq_min = -np.ones((4,1))
+dq_max = np.ones((4,1))
 
 # jacobian
-J = four_link_jacobian(q)
-# jacobian
-M = four_link_inertia(q)
+J = robot.jacobian(q)
+# calculate the velocity polytope
+v_poly = velocity_polytope(J, dq_min ,dq_max)
 
-# calculate the acceleration polytope
-
-acc_poly = capacity.acceleration_polytope(J, M, tau_min ,tau_max)
-
-# calculate the acceleration ellipsoid
-acc_ellisoid= capacity.acceleration_ellipsoid(J, M, tau_max)
+# calculate the velocity ellipsoid
+v_ellipsoid = velocity_ellipsoid(J, dq_max)
 
 # visualise polytope ellipsoid
-fig = plt.figure(13, figsize=[10,10])
+fig = plt.figure(14, figsize=[10,10])
+scale = 1/5
 
-scale = 1/100
+# plot the robot
+robot_position = robot.forward_kinematics(q) 
+robot.plot(plt, q)
 
-#plot the robot
-robot_position = four_link_forward_kinematics(q) 
-four_link_plot_robot(plt, q)
+# plot the polytope
+plot_polytope(plot=plt,
+              polytope=v_poly,
+              center=robot_position, 
+              face_color='lightsalmon', 
+              edge_color='orangered',
+              vertex_color='gray',
+              label='polytope', 
+              scale=scale)
 
-#plot the polytope
-visual.plot_polytope(plot=fig,
-                    polytope=acc_poly,
-                    center=robot_position, 
-                    face_color='lightsalmon', 
-                    edge_color='orangered',
-                    vertex_color='gray',
-                    label='polytope', 
-                    scale=scale)
 # plot ellipsoid
-visual.plot_ellipsoid(ellipsoid=acc_ellisoid, 
+plot_ellipsoid(ellipsoid=v_ellipsoid, 
                center=robot_position, 
-               plot=fig,
+               plot=plt, 
                label='ellipsoid', 
                edge_color='blue', 
-               alpha=1.0, 
+               alpha=1.0,
                scale=scale)
+
+plt.title("Velocity capacity")
 plt.grid()
-plt.legend()
 plt.axis('equal')
+plt.legend()
 plt.show()
