+++ examples/4dof_robot_examples/force_capacity.py
@@ -1,10 +1,14 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-import pycapacity.robot as capacity # robot capacity module
-import pycapacity.visual as visual # visualistion tools
+from pycapacity.robot import * # robot capacity module
+from pycapacity.visual import * # visualistion tools
 
-from four_link_utils import *
+# four link robot import
+from pycapacity.examples import FourLinkRobot
+
+# create the robot
+robot = FourLinkRobot()
 
 # joint positions q
 q  = np.random.rand(4)*np.pi/3*2-1
@@ -13,24 +17,23 @@ tau_min = -np.ones((4,1))
 tau_max = np.ones((4,1))
 
 # jacobian
-J = four_link_jacobian(q)
-# calculate the velocity polytope
-f_poly = capacity.force_polytope(J,tau_min,tau_max)
+J = robot.jacobian(q)
+# calculate the force polytope
+f_poly = force_polytope(J,tau_min,tau_max)
 
-# calculate the velocity ellipsoid
-f_ellipsoid = capacity.force_ellipsoid(J, tau_max)
+# calculate the force ellipsoid
+f_ellipsoid = force_ellipsoid(J, tau_max)
 
 # visualise polytope ellipsoid
 fig = plt.figure(12, figsize=[10,10])
-
 scale = 1/5
 
-#plot the robot
-robot_position = four_link_forward_kinematics(q) 
-four_link_plot_robot(plt, q)
+# plot the robot
+robot_position = robot.forward_kinematics(q) 
+robot.plot(plt, q)
 
-#plot the polytope
-visual.plot_polytope(plot=fig,
+# plot the polytope
+plot_polytope(plot=plt,
               polytope=f_poly,
               center=robot_position, 
               face_color='lightsalmon', 
@@ -40,14 +43,14 @@ visual.plot_polytope(plot=fig,
               scale=scale)
 
 # plot ellipsoid
-visual.plot_ellipsoid(ellipsoid=f_ellipsoid, 
+plot_ellipsoid(ellipsoid=f_ellipsoid, 
                center=robot_position, 
-               plot=fig, 
+               plot=plt, 
                label='ellipsoid', 
                edge_color='blue', 
                alpha=1.0,
                scale=scale)
-
+plt.title('Force capacity')
 plt.grid()
 plt.axis('equal')
 plt.legend()
