+++ examples/4dof_robot_examples/four_link_utils.py
@@ -1,46 +1,0 @@
-"""
-Four link planar robot example
-
-Direct kinematics and jacobian matrix calacualiton for the simple 4 dof plananr robot
-
-n=4 and m=2
-
-"""
-
-import numpy as np
-# jacobian function for four link planar robot
-def four_link_jacobian(joints):
-    sq1 = np.sin(joints[0])
-    sq12 = np.sin(joints[0] + joints[1])
-    sq123 = np.sin(joints[0] + joints[1] + joints[2])
-    sq1234 = np.sin(joints[0] + joints[1] + joints[2] + joints[3])
-    cq1 = np.cos(joints[0])
-    cq12 = np.cos(joints[0] + joints[1])
-    cq123 = np.cos(joints[0] + joints[1] + joints[2])
-    cq1234 = np.cos(joints[0] + joints[1] + joints[2] + joints[3])
-    return np.array([[0.5*cq1+0.5*cq12+0.5*cq123+0.3*cq1234, 0.5*cq12+0.5*cq123+0.3*cq1234, 0.5*cq123+0.7*cq1234, +0.3*cq1234], [-0.5*sq1-0.5*sq12-0.5*sq123-0.3*sq1234, -0.5*sq12-0.5*sq123-0.3*sq1234, -0.5*sq123-0.3*sq1234, -0.3*sq1234]])
-# inertia matrix of a four link planar robot
-def four_link_inertia(joints):
-    sq1 = np.sin(joints[1])
-    cq1 = np.cos(joints[1])
-    sq2 = np.sin(joints[2])
-    cq2 = np.cos(joints[2])
-    sq3 = np.sin(joints[3])
-    cq3 = np.cos(joints[3])
[…]
