+++ pycapacity/__init__.py
@@ -3,3 +3,4 @@ import pycapacity.human
 import pycapacity.robot
 import pycapacity.objects
 import pycapacity.visual
+import pycapacity.examples
