### 2023-07-24 9a00d33
added examples module and tests + docs for swig updated with absolute path + pinoscchio more details - issue #5
