### 2023-07-24 auctus-team/pycapacity#5 @askuric
Hey @sea-bass,

Thanks for the feedback.

## robotics toolbox

So first of all the Swift one is a tricky one, I am not sure what is going on. However, there seems to be an issue related to it:
https://github.com/jhavl/swift/issues/47

It seems to be something related to the new swig version.
However, I think I was able to find the issue. It seems that the local path to the generated stl file was not properly found. So it needs to be set to the absolute path.  I've updated the docs to include the absolute path.

> Also, I had to additionally pip install networkx because it was another missing dependency not listed anywhere.

Thanks, I've updated the docs to include all the missing packages.

## pinocchio

> The pip instructions just say pip install pin, but actually you need other things like example-robot-data, meshcat, etc. Make sure this is consistent.

Thanks, I've updated the docs to take this in consideration. I was completely concentrated on anaconda install that I've forgot to write the steps for the pip version.

> Also, the first (static) Meshcat example works for me but it's all blank... any ideas?

Here I think the issue is the PC speed, sometimes the pc does the rest of the program faster than the visualisation is opened and then some strange stuff can happen. Could you please try adding a timeout after `viz.initViewer(open=True)` and `viz.display(q0)` something like this:

```python
time.sleep(0.5)
```

Also you'll gona have to import the time module

```python
import time
```

If this works I'll add it directly to the library docs.

## four link robot

Four link robot is convenient for testing so I've put it into a module called examples. Maybe we're gonna have some more toy systems inside at some point. So with this in mind the code in the docs is much simpler and no need to create additionally files.  Here is the [module](https://auctus-team.github.io/pycapacity/pycapacity.examples.html) and here are the updated [docs](https://auctus-team.github.io/pycapacity/examples/fourlink.html) .
