### 2023-07-25 auctus-team/pycapacity#5 @askuric
Awesome,

I've added a delay in the [pinocchio example](https://auctus-team.github.io/pycapacity/examples/pinocchio.html), as well as a short comment saying why its there and what to do if the model still does not show.

> Unsure if you should do this, but fixing the random seed will ensure that the output shown in the doc is identical to that the users will see. Then, users can opt to either leave out the seed or change it.

Makes perfect sense, I've added a seed for all the randomly generated examples so that they initially correspond to the images in the docs.  For example in the case of 4 link [here](https://auctus-team.github.io/pycapacity/examples/fourlink.html) and random human [here](https://auctus-team.github.io/pycapacity/examples/randomised_human.html).
