### 2023-08-01 auctus-team/pycapacity#9 @JHartzer
[JOSS REVIEW] Additional Paper Feedback
# General Feedback
As mentioned in #6, I would suggest not saying the software is real-time as there are no guarantees that the software will produce calculations within a certain time frame. Moreover, it is technically possible for system interrupts to cause a python function call to take *much* longer than the times listed because of various OS functions. Because you mention safety applications in your paper, I would stress that this software is at best *near-real-time* or *online*. However, as you still would like to mention the performance, timing variance or distribution would be useful to add.

Additionally, I believe there should be a mention of the rate at which the functions are called. The execution times are useful to see, but if I need to call the Acceleration polytope metric more than 3 times per second, then this software is no longer even near real-time. It is my understanding that inverse kinematics and other constraints are calculated tens or hundreds of times per second and I wonder if this would be a limitation when considering musculoskeletal models.

There should be additional discussion and citations on the current state of the field in regards to alternative software solutions and where this work differs.

Lastly, the paper ends very abruptly. The addition of concluding remarks would help greatly as the introduction and statement of need are already strong.

## Introduction
- "where he lacks" use the "they"
- It can be interesting ~~to be able~~ to compute these metrics

## Statement of Need
- "real-time" should be updated to "near real-time" or "online"
- "aslo" -> "also"

## Iterative convex-hull method
- Remove "And" from the start of the third sentence and remove the paragraph break

## Robotic manipulator metrics
- "For ~~the~~ robotics manipulators, the package integrates"

## Polytopes
- Move "This approach is described by..." to the end of the following paragraph

## Polytope evaluation algorithms used
- Create captions for the performance tables
- Reference the table itself in the text, similar to how the figures were done
