### 2023-08-01 #5670 @JHartzer
I have complete my review of the software and paper and have a single blocking issue regarding the paper to ask for additional discussion on rates and real-time operation as well as general grammatical nitpicks
auctus-team/pycapacity#9
