### 2023-08-10 b7a9a61
refactored examples
