### 2023-08-10 b46c559
added readme in examples subfolders
