+++ examples/notebooks/README.md
@@ -1,0 +1,12 @@
+# Interactive Jupyter Notebooks
+
+The Jupyter notebooks are located in the `examples/notebooks` folder.
+Currently, the following notebooks are available where no dependencies are required:
+- [Four link robot](four_link.ipynb)
+- [Randomised robot and human model](demo_simple.ipynb)
+
+
+The other notebooks require additional dependencies:
+- [pinocchio](pinocchio.ipynb)
+- [robotics toolbox with swift](robotics_toolbox_swift.ipynb)
+- [robotics toolbox with pyplot](robotics_toolbox_pyplot.ipynb)
