+++ examples/scripts/README.md
@@ -1,0 +1,18 @@
+# Python Example Scripts
+
+The Python scripts are located in the `examples/scripts` folder.
+Currently, the following scripts are available where no dependencies are required:
+- **Four link robot**
+    - [Acceleration Polytope and Ellipsoid](scripts/4dof_robot_examples/acceleration_capacity.py)
+    - [Force Polytope and Ellipsoid](scripts/4dof_robot_examples/velocity_capacity.py)
+    - [Velocity Polytope and Ellipsoid](scripts/4dof_robot_examples/velocity_capacity.py)
+- Random robot model
+    - [Polytope](scripts/robot_random_model.py)
+    - [Comparison Polytope and Ellipsoid](scripts/robot_random_polytope_ellispoid.py)
+- [Random human model](scripts/human_random_model.py)
+
+The other scripts require additional dependencies:
+- [pinocchio](scripts/pinocchio.py)
+- [robotics toolbox with swift](scripts/robotics_toolbox_swift.py)
+- [robotics toolbox with pyplot](scripts/robotics_toolbox_pyplot.py)
+- [pyomeca](scripts/pyomeca.py)
