+++ examples/README.md
@@ -17,6 +17,23 @@ The other notebooks require additional dependencies:
 - [robotics toolbox with swift](notebooks/robotics_toolbox_swift.ipynb)
 - [robotics toolbox with pyplot](notebooks/robotics_toolbox_pyplot.ipynb)
 
+
+### Running the example notebooks
+
+Open the terminal in this folder `examples/notebooks` and if you do not have jupyter installed, do install it
+
+```
+pip install jupyter
+```
+or with anaconda
+```
+conda install -c anaconda-forge jupyter
+```
+Once you have jupyter installed just run and you'll be able to navigate the example scripts and run them
+```
+jupyter lab
+```
+
 ## Python scripts
 
 The Python scripts are located in the `examples/scripts` folder.
@@ -35,3 +52,11 @@ The other scripts require additional dependencies:
 - [robotics toolbox with swift](scripts/robotics_toolbox_swift.py)
 - [robotics toolbox with pyplot](scripts/robotics_toolbox_pyplot.py)
 - [pyomeca](scripts/pyomeca.py)
+
+
+### Running the example scripts
+
+Open the terminal in this folder `examples/scripts` and if all the dependencies are met you should be able to run the scripts with
+```
+python name_of_the_script.py
+```
