+++ examples/scripts/README.md
@@ -16,3 +16,10 @@ The other scripts require additional dependencies:
 - [robotics toolbox with swift](robotics_toolbox_swift.py)
 - [robotics toolbox with pyplot](robotics_toolbox_pyplot.py)
 - [pyomeca](pyomeca.py)
+
+## Running the example scripts
+
+Open the terminal in this folder `examples/scripts` and if all the dependencies are met you should be able to run the scripts with
+```
+python name_of_the_script.py
+```
