### 2023-08-10 8eb6bb8
readme examples how to run - typo
