+++ examples/scripts/README.md
@@ -18,7 +18,7 @@ The other scripts require additional dependencies:
 - [pyomeca](pyomeca.py)
 
 
-### Running the example scripts
+### Running the example scripts
 
 Open the terminal in this folder `examples/scripts` and if all the dependencies are met you should be able to run the scripts with
 ```
