+++ examples/README.md
@@ -55,7 +55,7 @@ The other scripts require additional dependencies:
 
 
 
-### Running the example scripts
+### Running the example scripts
 
 Open the terminal in this folder `examples/scripts` and if all the dependencies are met you should be able to run the scripts with
 ```
