### 2023-08-10 7ede719
added links to the examples folder in the docs, issue #4
