+++ docs/source/examples/randomised_human.md
@@ -4,6 +4,18 @@ Randomized Human musculoskeletal models
 A simple example program for force polytope evaluation of a randomised human musculoskeletal model. 
 Simply change the number of dof, number of forces and force limits and see how the calculation time and shape evaluates.
 
+
+<div class="admonition-new-examples admonition">
+<p class="admonition-title">📢 NEW Examples!</p>
+<dl class="simple">
+<dt>For some more examples check out the <code class="docutils literal notranslate"><span class="pre">examples</span></code> folder of the repository.</dt><dd><ul class="simple">
+<li><p>Interactive jupyter notebooks are available in the <code class="docutils literal notranslate"><span class="pre">examples/notebooks</span></code> folder: <a class="reference external" href="https://github.com/auctus-team/pycapacity/blob/master/examples/notebooks/demo_simple.ipynb">see on Github</a></p></li>
+</ul>
+</dd>
+</dl>
+</div>
+
+
 ```python
 import pycapacity.human as capacity # robot capacity module
 import numpy as np

[…]
