### 2023-08-10 auctus-team/pycapacity#4 @askuric
Hey @sea-bass,

Sorry for the delay, I've not had much time to spend on pycapacity.
Thank's for the suggestion, its a good idea. I did not do it at the start because most of these libraries require adding more dependencies.  But I've took your suggestion and I've documented the examples a bit more than before and pointed to the docs page to find the full tutorial, so it should not be too hard to use.

So, I've added some more examples in the examples folder and separated them in the notebooks and scripts.
The notebooks are really cool, as they are interactive and you can see the polytopes and ellipsoids visualised as you move the robot :D
I've also added the links in the docs to direct people to the examples.
