### 2023-08-10 49851b9
paper changes + some state of the art
