+++ paper/paper.bib
@@ -147,3 +147,101 @@ doi={10.1109/ICRA48506.2021.9561366}}
   year={2009},
   organization={Kobe, Japan}
 }
+
+@article{Haviland2020Maximising,
+  author       = {Jesse Haviland and
+                  Peter Corke},
+  title        = {Maximising Manipulability During Resolved-Rate Motion Control},
+  journal      = {CoRR},
+  volume       = {abs/2002.11901},
+  year         = {2020},
+  url          = {https://arxiv.org/abs/2002.11901},
+  eprinttype    = {arXiv},
+  eprint       = {2002.11901},
+  timestamp    = {Tue, 03 Mar 2020 14:32:13 +0100},
+  biburl       = {https://dblp.org/rec/journals/corr/abs-2002-11901.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
+}
+
+@inproceedings{laisne2023genetic,
+  TITLE = {{Genetic algorithms for force polytopes prediction}},
+  AUTHOR = {Laisn{\'e}, Gautier and Salotti, Jean-Marc and Rezzoug, Nasser},
+  URL = {https://inria.hal.science/hal-04151258},
+  BOOKTITLE = {{48{\`e}me Congr{\`e}s de la Soci{\'e}t{\'e} de Biom{\'e}canique}},
+  ADDRESS = {Grenoble, France},
+  ORGANIZATION = {{Soci{\'e}t{\'e} de Biom{\'e}canique}},
+  YEAR = {2023},
+  MONTH = Oct,
+  KEYWORDS = {genetic algorithm ; musculoskeletal model ; force polytope},
+  HAL_ID = {hal-04151258},
+  HAL_VERSION = {v1},
+}
+@misc{pypoman,
+  author = {Caron, S.},
+  title = {Pypoman: Polyhedron manipulation in Python},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/stephane-caron/pypoman}}
+}
+@inproceedings{mpt3,
+  author={Herceg, Martin and Kvasnica, Michal and Jones, Colin N. and Morari, Manfred},
+  booktitle={2013 European Control Conference (ECC)}, 
+  title={Multi-Parametric Toolbox 3.0}, 
+  url="https://www.mpt3.org/",
+  year={2013},
+  volume={},
+  number={},
+  pages={502-510},
+  doi={10.23919/ECC.2013.6669862}
+}
+
+@misc{manipulability_metrics,
+  author = {Prada, Miguel},
+  title = {Manipulability Metrics: ROS-based library for computing manipulability metrics},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/tecnalia-medical-robotics/manipulability_metrics}}
+}
+
+
+@misc{constrained_manipulability,
+  author = {Long, Philip},
+  title = {Constrained Manipulability: Repository containing the codes for manipulability learning, tracking and transfer.},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/NoemieJaquier/Manipulability}}
+}
+@INPROCEEDINGS{Long2018Evaluating,
+  author={P. {Long} and T. {Padir}},
+  booktitle={2018 IEEE-RAS 18th International Conference on Humanoid Robots (Humanoids)},
+  title={Evaluating Robot Manipulability in Constrained Environments by Velocity Polytope Reduction},
+  year={2018},
+  volume={},
+  number={},
[…]
