+++ paper/paper.bib
@@ -208,12 +208,13 @@ doi={10.1109/ICRA48506.2021.9561366}}
 
 @misc{constrained_manipulability,
   author = {Long, Philip},
-  title = {Constrained Manipulability: Repository containing the codes for manipulability learning, tracking and transfer.},
+  title = {Constrained Manipulability: A library used to compute and visualize a robot's capacities in constrained environments.},
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
-  howpublished = {\url{https://github.com/NoemieJaquier/Manipulability}}
+  howpublished = {\url{https://github.com/philip-long/constrained_manipulability}}
 }
+
 @INPROCEEDINGS{Long2018Evaluating,
   author={P. {Long} and T. {Padir}},
   booktitle={2018 IEEE-RAS 18th International Conference on Humanoid Robots (Humanoids)},
@@ -221,14 +222,14 @@ doi={10.1109/ICRA48506.2021.9561366}}
   year={2018},
   volume={},
   number={},
-  url={https://github.com/philip-long/constrained_manipulability}
   pages={1-9},
   doi={10.1109/HUMANOIDS.2018.8624962},
   ISSN={2164-0580},
-  month={Nov},}
+  month={Nov},
+}
 
 
-@misc{manipulability_metrics,
+@misc{manipulability,
   author = {Jaquier, Noémie},
   title = {Manipulability: Repository containing the codes for manipulability learning, tracking and transfer.},
   year = {2023},
