+++ paper/paper.bib
@@ -1,248 +1,222 @@
-@INPROCEEDINGS{Skuric2021,
-  author={Skuric, Antun and Padois, Vincent and Daney, David},
-  booktitle={2021 IEEE International Conference on Robotics and Automation (ICRA)}, 
-  title={On-line force capability evaluation based on efficient polytope vertex search}, 
-  year={2021},
-  volume={},
-  number={},
-  pages={1700-1706},
-  doi={10.1109/ICRA48506.2021.9562050}
+@inproceedings{Skuric2021,
+	title        = {On-line force capability evaluation based on efficient polytope vertex search},
+	author       = {Skuric, Antun and Padois, Vincent and Daney, David},
+	year         = 2021,
+	booktitle    = {2021 IEEE International Conference on Robotics and Automation (ICRA)},
+	volume       = {},
+	number       = {},
+	pages        = {1700--1706},
+	doi          = {10.1109/ICRA48506.2021.9562050}
 }
-
-@ARTICLE{Skuric2022,
-  author={Skuric, Antun and Padois, Vincent and Rezzoug, Nasser and Daney, David},
-  journal={IEEE Robotics and Automation Letters}, 
-  title={On-Line Feasible Wrench Polytope Evaluation Based on Human Musculoskeletal Models: An Iterative Convex Hull Method}, 
-  year={2022},
-  volume={7},
-  number={2},
-  pages={5206-5213},
-  doi={10.1109/LRA.2022.3155374}
+@article{Skuric2022,
+	title        = {On-Line Feasible Wrench Polytope Evaluation Based on Human Musculoskeletal Models: An Iterative Convex Hull Method},
+	author       = {Skuric, Antun and Padois, Vincent and Rezzoug, Nasser and Daney, David},
+	year         = 2022,
+	journal      = {IEEE Robotics and Automation Letters},
+	volume       = 7,
+	number       = 2,
+	pages        = {5206--5213},
+	doi          = {10.1109/LRA.2022.3155374}
 }
-
-@InProceedings{Gouttefarde2010,
-author="Gouttefarde, M.
-and Krut, S.",
-editor="Lenarcic, Jadran
-and Stanisic, Michael M.",
-title="Characterization of Parallel Manipulator Available Wrench Set Facets",
-booktitle="Advances in Robot Kinematics: Motion in Man and Machine",
-year="2010",
-publisher="Springer Netherlands",
-address="Dordrecht",
-pages="475--482",
-abstract="The available wrench set is the set of wrenches that can be generated at the platform of a parallel manipulator. In a given configuration, this set is known to be a convex polytope and, recently, it has been pointed out that it is in fact a particular type of convex polytope called a zonotope. This paper deals with the case of parallel manipulators having as many or more actuators than degrees of freedom and discusses a characterization of the facets of the available wrench set.",
-isbn="978-90-481-9262-5"
+@inproceedings{Gouttefarde2010,
+	title        = {Characterization of Parallel Manipulator Available Wrench Set Facets},
+	author       = {Gouttefarde, M. and Krut, S.},
+	year         = 2010,
+	booktitle    = {Advances in Robot Kinematics: Motion in Man and Machine},
+	publisher    = {Springer Netherlands},
+	address      = {Dordrecht},
+	pages        = {475--482},
+	isbn         = {978-90-481-9262-5},
+	editor       = {Lenarcic, Jadran and Stanisic, Michael M.},
[…]
