+++ paper/paper.bib
@@ -220,3 +220,40 @@
 	number       = {2-3},
 	pages        = {624--650}
 }
+
+
+@misc{pygradientpolytope,
+	title        = {pygradientpolytope: A set of ROS nodes for generating and visualizing Cartesian Velocity, Force and Desired polytopes},
+	author       = {Sagar, Keerthi},
+	year         = 2023,
+	journal      = {GitLab repository},
+	publisher    = {GitLab},
+	howpublished = {\url{https://gitlab.com/KeerthiSagarSN/rospygradientpolytope}}
+}
+
+@article{fukuda2004frequently,
+  title={Frequently asked questions in polyhedral computation},
+  author={Fukuda, Komei and others},
+  journal={ETH, Zurich, Switzerland},
+  volume={85},
+  year={2004}
+}
+
+
+
+@misc{cddlib,
+	title        = {cddlib: An efficient implementation of the Double Description Method},
+	author       = {Sagar, Keerthi},
+	year         = 2023,
+	journal      = {GitHub repository},
+	publisher    = {GitHub},
+	howpublished = {\url{https://github.com/cddlib}}
+}
+
+@article{fukuda1997cdd,
+  title={cdd/cdd+ Reference Manual},
+  author={ETH Zurich},
+  journal={Institute for Operations Research, ETH-Zentrum},
+  pages={91--111},
+  year={1997}
+}
