### 2023-08-11 4bc731d
bold symbol try
