+++ paper/paper.bib
@@ -243,7 +243,7 @@
 
 @misc{cddlib,
 	title        = {cddlib: An efficient implementation of the Double Description Method},
-	author       = {Sagar, Keerthi},
+	author={ETH Zurich},
 	year         = 2023,
 	journal      = {GitHub repository},
 	publisher    = {GitHub},
@@ -252,7 +252,7 @@
 
 @article{fukuda1997cdd,
   title={cdd/cdd+ Reference Manual},
-  author={ETH Zurich},
+  author={Fukuda, Komei},
   journal={Institute for Operations Research, ETH-Zentrum},
   pages={91--111},
   year={1997}
