### 2023-08-11 de3710c
added discussion about polytopes and ellipsoids
