+++ paper/paper.md
@@ -38,7 +38,7 @@ Beyond the characterization of human physical capabilities, as today's collabora
 
 There are many different metrics available in the literature that might be used to characterize physical abilities: force capacity, velocity capacity, acceleration capacity, accuracy, stiffness etc. Most of these metrics can be represented by two families of geometric shapes: ellipsoids [@yoshikawa1985manipulability] and polytopes [@chiacchio1997force]. These metrics are traditionally important tools for off-line analysis purposes (workspace design, human motion and ergonomics analysis) and recently, they have shown a great potential to be used for interactive online applications, to be integrated in robot control strategies or as a visual feedback to the operator. 
 
-Ellipsoid metrics are often used for evaluating the manipulability of the robot's end-effector. The manipulability ellipsoid is a geometric shape that represents the robot's ability to move with in the task space. The manipulability ellipsoid is defined by its principal axis that can be found very efficiently using the singular value decomposition (SVD) of the robot's Jacobian matrix [@yoshikawa1985manipulability]. Due to their computational efficiency and intuitive visualisation, they have been used in many different applications, such as robot control, workspace design, robot design, etc. Therefore there are several open-source packages that implement the manipulability ellipsoid evaluation and visualisation, such as `MMC` [@Haviland2020Maximising], `manipulability_metrics` [@manipulability_metrics], `Manipulability` [@manipulability;Jaquier2021Geometry].However, all of these packages are limited to the evaluation of the manipulability ellipsoid, representing the velocity capacity, and they do not provide tools for evaluating other ellipsoid metrics, such as force capacity, acceleration capacity, etc. Additionally these software packages are often developed for the use with a specific robotics library, such as `robotic-toolbox` [@corke2021not] or `ROS` [@carpentier2019pinocchio], and they are not trivial to integrate with other libraries.
+Ellipsoid metrics are often used for evaluating the manipulability of the robot's end-effector. The manipulability ellipsoid is a geometric shape that represents the robot's ability to move with in the task space. The manipulability ellipsoid is defined by its principal axis that can be found very efficiently using the singular value decomposition (SVD) of the robot's Jacobian matrix [@yoshikawa1985manipulability]. Due to their computational efficiency and intuitive visualisation, they have been used in many different applications, such as robot control, workspace design, robot design, etc. Therefore there are several open-source packages that implement the manipulability ellipsoid evaluation and visualisation, such as `MMC` [@Haviland2020Maximising], `manipulability_metrics` [@manipulability_metrics], `Manipulability` [@manipulability;@Jaquier2021Geometry].However, all of these packages are limited to the evaluation of the manipulability ellipsoid, representing the velocity capacity, and they do not provide tools for evaluating other ellipsoid metrics, such as force capacity, acceleration capacity, etc. Additionally these software packages are often developed for the use with a specific robotics library, such as `robotic-toolbox` [@corke2021not] or `ROS` [@carpentier2019pinocchio], and they are not trivial to integrate with other libraries.
 
[…]
