+++ paper/paper.md
@@ -48,8 +48,7 @@ Therefore, this paper presents a Python `pycapacity` package in an effort to pro
 
 `pycapacity` has been used in several scientific papers, for real-time control of collaborative carrying using two Franka Emika Panda robots [@Skuric2021], for developing an assist-as-needed control strategy for collaborative carrying task of the human operator and the Franka robot [@Skuric2022]. The package has also been used to calculate the approximation of the robot's reachable space using convex polytope [@skuric2023]. On the other hand, the package has been used for the biomechanical calibration of the human musculoskeletal model [@laisne2023genetic].
 
-# Ellipsoids and polytopes as physical ability metrics
-
+# Ellipsoids and polytopes as physical ability metrics
 
 In robotics, different task space physical ability metrics establish the relationship between different limits of robot's actuators (joint positions, velocities, torques, etc.), its kinematics and dynamics, and the achievable sets of different task related physical quantities, such as achievable positions, velocities, forces and similar. Similar metrics can be established for humans as well, by leveraging their musculoskeletal models. Where the humans in addition to the joint limits (joint positions and velocities) have additional limits due to their using their muscles as actuators (contraction forces and velocities).
 
@@ -76,13 +75,13 @@ Figure 2. shows the difference between the force ellipsoid and polytope [@chiacc
 
 Ellipsoids, however, are much more present in the literature, as their computation is much faster than the computation of polytopes. 
 
-## Evaluating ellipsoids
+# Evaluating ellipsoids
 
 Evaluating ellipsoids is a computationally efficient task, as it can be done using the singular value decomposition (SVD) [@yoshikawa1985manipulability]. Ellipsoids can be fully defined using its principal axis and principal axis lengths. Once they are known, the ellipsoid can be easily visualized and used for further analysis.
 
 This package provides tools for evaluating several common ellipsoid metrics for robots and humans, such as velocity (manipulability), force and acceleration, and it provides a set of tools for their easy visualization. 
 
-## Evaluating polytopes
+## Evaluating polytopes
 
 Evaluating polytopes consists in finding either the minimal set of their vertices, $\mathcal{V}$-representation, or the minimal set of the half-planes defining their faces , $\mathcal{H}$-representation. The $\mathcal{V}$-representation is often used for visualization purposes, while the $\mathcal{H}$-representation is often integrated in different optimization problems, as it can be represented as a set of linear inequalities.
 
@@ -91,54 +90,6 @@ However, finding the $\mathcal{V}$-representation or the $\mathcal{H}$-represent
 Therefore, polytope evaluation is often a bottleneck in the computation of different physical ability metrics, especially for human musculoskeletal models, which have a large number of degrees of freedom and a large number of muscles. Furthermore, due to the inherent complexity of the polytope evaluation algorithms, finding the appropriate algorithm for a given polytope formulation and dimensionality of the input and output spaces is not a trivial task.
 
 This package aims to provide a selection of algorithms for polytope evaluation, capable of evaluating common physical ability polytopes in an easy to use and efficient way. These algorithms are implemented in Python and can be used as standalone tools as well. Additionally, the package provides tools for easy visualization the 2D and 3D polytopes.
-
[…]
