### 2023-08-11 8e01e57
added more in depth exection time
