+++ paper/paper.md
@@ -230,7 +230,7 @@ The methods for resolution of the polytope based metrics depend on the family of
 
 
 Polytope Metric | Algorithm | Problem type | Execution time <br>(mean $\pm$ std) [ms]
---- | -- | ---- | ----
+--- | -- | ----- | ----
 Velocity | HPSM | $x=By,~~ y_{min} \leq y \leq y_{max}$ | 5.2ms $\pm$ 1.44 [11.7]
 Acceleration |  HPSM | $x=By,~~ y_{min} \leq y \leq y_{max}$ | 4.3ms $\pm$ 0.43 [7.97]
 Force  | VEPOLI$^2$ | $Ax=b, ~~ b_{min} \leq b \leq y_{max}$| 6.2ms $\pm$ 1.42 [15.9]
