+++ paper/paper.md
@@ -231,22 +231,22 @@ The methods for resolution of the polytope based metrics depend on the family of
 
 Polytope Metric | Algorithm | Problem type | Execution time (mean $\pm$ std) [ms]
 --- | -- | ----- | ------
-Velocity | HPSM | $x=By,~~ y_{min} \leq y \leq y_{max}$ | 5.2ms $\pm$ 1.44 [11.7]
-Acceleration |  HPSM | $x=By,~~ y_{min} \leq y \leq y_{max}$ | 4.3ms $\pm$ 0.43 [7.97]
-Force  | VEPOLI$^2$ | $Ax=b, ~~ b_{min} \leq b \leq y_{max}$| 6.2ms $\pm$ 1.42 [15.9]
-Force intersection |  VEPOLI$^2$ | $Ax=b,~~ b_{min} \leq b \leq b_{max}$ | 83.9ms $\pm$ 28.4 [172.2]
-Force sum |  VEPOLI$^2$ | $Ax=b,~~ b_{min} \leq b \leq b_{max}$ | 18.1ms $\pm$ 6.4 [45.9]
-Reachable space |  ICHM | $x=By,~~  y \in P_{y}$ | 125ms $\pm$ 25.9 [191.2]
+Velocity | HPSM | $x=By,~ y \in [y_{min}, y_{max}]$ | 5.2ms $\pm$ 1.44 [11.7]
+Acceleration |  HPSM | $x=By,~ y \in [y_{min}, y_{max}]$ | 4.3ms $\pm$ 0.43 [7.97]
+Force  | VEPOLI$^2$ | $Ax=b, ~ y \in [y_{min}, y_{max}]$| 6.2ms $\pm$ 1.42 [15.9]
+Force intersection |  VEPOLI$^2$ | $Ax=b,~ b \in [b_{min}, b_{max}]$ | 83.9ms $\pm$ 28.4 [172.2]
+Force sum |  VEPOLI$^2$ | $Ax=b,~ b \in [b_{min}, b_{max}]$ | 18.1ms $\pm$ 6.4 [45.9]
+Reachable space |  ICHM | $x=By,~  y \in P_{y}$ | 125ms $\pm$ 25.9 [191.2]
 
 The average execution time is calculated using 7 dof Franka Emika panda robot, the model was used with `pinocchio` software. All the experiments are run on a computer equipped with 1.90GHz Intel i7-8650U processor.
 
 In case of human musculoskeletal models the methods used are given in the table below.
 
-Polytope Metric  | Algorithm | Problem type | Execution time 
+Polytope Metric  | Algorithm | Problem type | Execution time (mean $\pm$ std) [ms]
 -- | --- | ----- | ---
-Force  | ICHM | $Ax=By,~~ y_{min} \leq y \leq y_{max}$ | ~ 200ms 
-Acceleration |  HPSM or ICHM | $x=By,~~ y_{min} \leq y \leq y_{max}$ |  ~ 300ms 
-Velocity | ICHM | $x=By,~~ y \in P_{y}$ | ~ 200ms
+Force  | ICHM | $Ax=By,~ y \in [y_{min}, y_{max}]$ | 18.1ms $\pm$ 1.2 [22.4]
+Acceleration |  HPSM or ICHM | $x=By,~ y \in [y_{min}, y_{max}]$ |  388.4ms $\pm$ 51.2 [666.0]
+Velocity | ICHM | $x=By,~ y \in P_{y}$ | 18.1ms $\pm$ 1.2 [22.4]
 
 The average execution time was calculated using 50 muscle 7 dof musculoskeletal model introduced by [@holzbaur2005model], the model was used with `biorbd` biomechanics software. The experiments are run on a computer equipped with 1.90GHz Intel i7-8650U processor.
 
