+++ paper/paper.md
@@ -230,7 +230,7 @@ The methods for resolution of the polytope based metrics depend on the family of
 
 
 Polytope Metric | Algorithm | Problem type | Execution time <br> mean $\pm$ std. (max) [ms]
---- | -- | ----- | -----
+--- | -- | ----- | ----
 Velocity | HPSM | $x=By,~ y \in [y_{min}, y_{max}]$ | 3.6 $\pm$ 0.21 (5.7)
 Acceleration |  HPSM | $x=By,~ y \in [y_{min}, y_{max}]$ | 6.6 $\pm$ 1.4 (14.2)
 Force  | VEPOLI$^2$ | $Ax=b, ~ y \in [y_{min}, y_{max}]$| 6.8 $\pm$ 0.88 (16.4)
@@ -244,9 +244,9 @@ In case of human musculoskeletal models the methods used are given in the table 
 
 Polytope Metric  | Algorithm | Problem type | Execution time <br> mean $\pm$ std. (max) [ms]
 -- | --- | ----- | ---
-Force  | ICHM | $Ax=By,~ y \in [y_{min}, y_{max}]$ | 18.1 $\pm$ 1.2 (22.4)
-Acceleration |  HPSM or ICHM | $x=By,~ y \in [y_{min}, y_{max}]$ |  388.4 $\pm$ 51.2 (666.0)
-Velocity | ICHM | $x=By,~ y \in P_{y}$ | 18.1 $\pm$ 1.2 (22.4)
+Force  | ICHM | $Ax=By,~ y \in [y_{min}, y_{max}]$ | 186.8 $\pm$ 45.6 (281.6)
+Acceleration |  HPSM or ICHM | $x=By,~ y \in [y_{min}, y_{max}]$ |  378.8 $\pm$ 62.3 (643.7)
+Velocity | ICHM | $x=By,~ y \in P_{y}$ | 223.1 $\pm$ 60.4 (389.1)
 
 The average execution time was calculated using 50 muscle 7 dof musculoskeletal model introduced by [@holzbaur2005model], the model was used with `biorbd` biomechanics software. The experiments are run on a computer equipped with 1.90GHz Intel i7-8650U processor.
 
