+++ examples/scripts/benchmarking/README.md
@@ -4,9 +4,9 @@
 Performance testing and benchmarking scripts available:
 
 No dependencies required:
-- [Performance testing script for polytope algorithms using radom robot models](scripts/benchmarking/polytope_robot_performance_analysis.py)
-- [Performance testing script for polytope algorithms using radom human models](scripts/benchmarking/polytope_human_performance_analysis.py)
+- [Performance testing script for polytope algorithms using radom robot models](polytope_robot_performance_analysis.py)
+- [Performance testing script for polytope algorithms using radom human models](polytope_human_performance_analysis.py)
 
 Dependencies required:
-- [Performance testing script for polytope algorithms using pinocchio](scripts/benchmarking/polytope_robot_performance_analysis_pinocchio.py)
-- [Performance testing script for polytope algorithms using biorbd](scripts/benchmarking/polytope_human_performance_analysis_biorbd.py)
+- [Performance testing script for polytope algorithms using pinocchio](polytope_robot_performance_analysis_pinocchio.py)
+- [Performance testing script for polytope algorithms using biorbd](polytope_human_performance_analysis_biorbd.py)
