### 2023-08-11 b7188bb
added performance to docs
