+++ docs/source/algorithms.rst
@@ -1,0 +1,141 @@
+
+Polytope evaluation algorithms
+------------------------------
+There are three methods implemented in this paper to resolve all the polytope calculations:
+
+* Hyper-plane shifting method (HPSM)
+* Iterative convex hull method (ICHM)
+* Vertex enumeration (VEPOLI\ :sup:`2`)
+
+All of the methods are implemented in the module `pycapacity.algorithms` and can be used as standalone functions.  See in [docs for more info](pycapacity.algorithms.html). 
+
+Hyper-plane shifting method (HPSM)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+| Based on the paper:
+| `Characterization of Parallel Manipulator Available Wrench Set Facets <http://www.lirmm.fr/krut/pdf/2010_gouttefarde_ark-0602650368/2010_gouttefarde_ark.pdf>`_
+| by Gouttefarde M., Krut S. 
+| In: Lenarcic J., Stanisic M. Advances in Robot Kinematics: Motion in Man and Machine. Springer, Dordrecht (2010)
+
+This method finds the half-space representation of the polytope of a class:
+
+.. math:: P = \{ x ~|~ x = By, \quad y_{min}\leq y \leq y_{max} \}
+
+To find the vertices of the polytope after finding the half-space representation :math:`Hx \leq d` an convex-hull algorithm is used. 
+
+The method is a part of the ``pycapacity.algorithms``` module ``hyper_plane_shift_method``, See in `docs for more info <pycapacity.algorithms.html#pycapacity.algorithms.hyper_plane_shift_method>`_ 
+
+Iterative convex-hull method (ICHM)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+| Based on the paper:
+| `On-line feasible wrench polytope evaluation based on human musculoskeletal models: an iterative convex hull method <https://hal.inria.fr/hal-03369576>`_ 
+| by A.Skuric, V.Padois, N.Rezzoug and D.Daney
+| Published in RAL & ICRA2022 
+
+This method finds both vertex and half-space representation of the class of polytopes:
+
+
+.. math:: P = \{ x ~|~ Ax = By, \quad y_{min}\leq y \leq y_{max} \}
+
+
+And it can be additionally extended to the case where there is an additional projection matrix $P$ making a class of problems:
+
+.. math:: P = \{ x ~|~ x= Pz, Az = By, \quad y_{min}\leq y \leq y_{max} \}
+
+
+The method is a part of the ``pycapacity.algorithms`` module ``iterative_convex_hull_method``. See the `docs for more info <pycapacity.algorithms.html#pycapacity.algorithms.iterative_convex_hull_method>`_
+
+Vertex enumeration algorithm (VEPOLI\ :sup:`2`)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+| Based on the paper:
+| `On-line force capability evaluation based on efficient polytope vertex search <https://arxiv.org/abs/2011.05226>`_ 
+| by A.Skuric, V.Padois and D.Daney
+| Published on ICRA2021
+
+This method finds vertex representation of the class of polytopes:
+
+.. math:: P = \{ x ~|~ Ax = y, \quad y_{min}\leq y \leq y_{max} \}
+
+
+To find the half-space representation (faces) of the polytope after finding the vertex representation  an convex-hull algorithm is used. 
+
[…]
