### 2023-08-11 cd60c68
paper update and typos
