### 2023-08-11 1230c77
added perfs to the docs and readme refactor
