+++ docs/source/README.rst
@@ -1,7 +1,7 @@
 About pycapacity
 ################
 
-Real-time capable task-space capacity calculation python pip package
+An efficient task-space capacity calculation package for robotics and biomechanics
 
 
 .. Note::
@@ -121,6 +121,89 @@ For the human musculoskeletal models this package implements the polytope metric
     .. math:: P_{f} = \{f ~| J^Tf = NF,~ F_{min}\leq F\leq F_{max} \}
 
 Where :math:`J` is the model's jacobian matrix, :math:`L` si the muscle length jacobian matrix, :math:`N= -L^T` is the moment arm matrix, :math:`f` is the vector of cartesian forces, :math:`$\dot{x}` and :math:`\ddot{x}` are vectors fo cartesian velocities and accretions, :math:`\dot{q}` is the vector of the joint velocities, :math:`\tau` is the vector of joint torques, :math:`\dot{l}` is the vector of the muscle stretching velocities and :math:`F` is the vector of muscular forces. 
+
+
+
+
+Performance evaluation of polytope metrics
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+The applicable methods to evaluate different polytope based metrics depend on the family of problems they correspond to. 
+Therefore this section brings the information about which algorithm is used for which polytope 
+metric and provides a brief performance evaluation their execution times.
+
+.. list-table:: Title
+    :widths: 25 25 50 50
+    :header-rows: 1
+
+    * - Polytope Metric
+      - Algorithm
+      - Problem type
+      - Execution time [ms]  mean  :math:`\pm` std. (max)
+    * - Velocity 
+      - HPSM 
+      - :math:`x=By,~ y \in [y_{min}, y_{max}]`
+      - 3.6 :math:`\pm` 0.21 (5.7)
+    * - Acceleration 
+      -  HPSM 
+      - :math:`x=By,~ y \in [y_{min}, y_{max}]`
+      - 6.6 :math:`\pm` 1.4 (14.2)
+    * - Force  
+      - VEPOLI$^2$ 
+      - :math:`Ax=b, ~ b \in [b_{min}, b_{max}]`
+      - 6.8 :math:`\pm` 0.88 (16.4)
+    * - Force intersection 
+      -  VEPOLI$^2$ 
+      - :math:`Ax=b,~ b \in [b_{min}, b_{max}]`
+      - 98.2 :math:`\pm` 29.33 (165.8)
+    * - Force sum 
+      -  VEPOLI$^2$ 
+      - :math:`Ax=b,~ b \in [b_{min}, b_{max}]` 
+      - 17.1 :math:`\pm` 3.4 (44.9)
+    * - Reachable space 
+      -  ICHM 
+      - :math:`x=By,~  y \in P_{y}`
+      - 30.5 :math:`\pm` 6.6 (76.7)
+
+The average execution time is calculated using 7 dof Franka Emika panda robot, the model was used with ``pinocchio`` software. 
+All the experiments are run on a computer equipped with 1.90GHz Intel i7-8650U processor. The results are obtained using 
+the benchmarking script provided in the ``examples`` folder, `script link <https://github.com/auctus-team/pycapacity/blob/master/examples/scripts/benchmarking/polytope_robot_performance_analysis_pinocchio.py>`_.
+
+
+In case of human musculoskeletal models the methods used are given in the table below.
+
+
+.. list-table:: Title
+    :widths: 25 25 50 50
+    :header-rows: 1
+
+    * - Polytope Metric
+      - Algorithm
+      - Problem type
[…]
