### 2023-08-11 bdc0e6a
added perfs to the docs and readme refactor 2
