### 2023-08-11 9ce6b80
added perfs to the docs and readme refactor 3
