### 2023-08-11 4d9b47c
added a table with libraries that pycapacity can be integrated with
