### 2023-08-11 auctus-team/pycapacity#6 @askuric
Hi  @sea-bass

Thanks for the detailed review I appreciate it.
I helped me a lot to go a step further and I ma pleased with the result.

The paper is almost entirely rewritten, but the main changes are:

- a bit more consequential Statement of need with a discussion about the existing solutions
- a bit more info about what is the main aim of the library (apart form the near real-time)
- I've done a bit more benchmarking and updated the table with the statistics, as well as provided the benchmarking scripts in the examples folder so that everyone can check on their system
- Added the conclusion
- A section about the differences of ellipsoids and polytopes, and what do they actually represent added

In terms of the real-time claim.

It is absolutely true that if the real-time would have been the only goal the python implementation does not make sense. The goal of the library is to provide efficient tools for their evaluation in a easy-to-use package. However, looking at the efficiency of the implemented algorithms , even though they are implemented in python, they are in many cases good enough for interactive applications.

Especially in terms of musculoskeletal models, they are usually restricted to off-line use, due to their inherent dimensionality issues (large number of muscles) . The ICHM algorithm implemented in this package is able however to evaluate their polytopes within a 200-300ms which might have taken hours with different methods.
So its much closer to real-time than before, but it is far from sub-millisecond that you'd probably expect form "real-time" claims.

However just as an example, we have used this package (both for robots and humans) for the real-time robot control in our lab, we ran the code at the frequency of 50 to 100Hz which is perfectly enough for many use-cases.

Finally, the area of using  task-space physical ability metrics (especially polytopes) in real time applications is very small at the moment, mostly due to their computational complexity, so having the real-time keyword in the title allows to reach further and show to the community that it is not as far from the real-time as they might think. :D
