### 2023-08-11 auctus-team/pycapacity#6 @ShravanTata
@askuric I have the same concern as @sea-bass on the real-time claim. I do agree with your rationale but may be think about replacing "Real-time" with "Fast" or "Efficient" keywords instead.

I would also suggest to add some of the numbers you mentioned above in the paper and may be an example where the computation is profiled
