### 2023-08-11 #5670 @ShravanTata
I have completed my review. Excellent work! Here are some of my minor feedback:

*Installation:* Does installation proceed as outlined in the documentation?
  + Everything works smoothly. When I tested in a clean virtual environment `matplotlib` there were no backend installed by default.
    I had to install one of the following packages `PyQt6, PySide6, PyQt5, PySide2`.
    It could be considered to make one of these as optional dependency if none are present.
    This does not impact the decision on paper acceptance.

*Performance:* If there are any performance claims of the software, have they been confirmed?
  (If there are no claims, please check off this item.)
  + I agree with @sea-bass  and @JHartzer that it would be good have the algorithm profiled and have some metrics for the real-time claim of the algorithms.
  + Left a comment here: https://github.com/auctus-team/pycapacity/issues/6#issuecomment-1675205604

 *Community guidelines:* Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
  + It would be great to have these [guidelines](https://auctus-team.github.io/pycapacity/contirbuting.html) under the main repo as well.

I have the same feedback on text as @sea-bass. Thank you
