### 2023-08-11 auctus-team/pycapacity#6 @ShravanTata
* Typo in line 21: "off line" --> "offline"/"off-line"
