+++ .gitlab-ci.yml
@@ -5,6 +5,7 @@ pages:
   tags:
   - ci.inria.fr
   script:
+  - pip install --upgrade pip setuptools wheel
   - pip install .
   - cd docs
   - pip install -r requirements.txt
