+++ .gitlab-ci.yml
@@ -5,7 +5,7 @@ pages:
   tags:
   - ci.inria.fr
   script:
-  - apt-get install gcc
+  - apt install build-essential
   - pip install .
   - cd docs
   - pip install -r requirements.txt
