### 2023-08-13 auctus-team/pycapacity#6 @sea-bass
Thanks @askuric -- took a look at your reworked paper and it looks awesome! Much clearer and gentler introduction to the readers.

My only "real" comment is that it's probably risky for the paper to provide direct links to the performance benchmarking scripts since this might change in the future. I'd just say "provided in the repository" and add instructions in the repo itself for reproducing the benchmarking for both the robot and musculoskeletal models.

Finally, some small editorial comments as I read through. Nothing major:

* In the summary, and again in the statement of need, you mention `robotic-toolbox` but I think they call the module `roboticstoolbox`. Either use this, or the full name "Robotics Toolbox for Python".
* In the summary, line 15, small typo "biomechanics softwares `opensim`" --> "biomechanics *software* `opensim`"
* Typo in line 57-58: "as they are essentially a triangulated meshes" --> "as they are essentially triangulated meshes"
* Line 65 missing a space between `cddlib` and the citation, same with line 69 in `pygradientpolytope` and its citation.
* Typo in like 75: "Futhermore" -> "Furthermore"
* Line 129: Extra space after the comma in "their faces, H-representation".
* Lines 208-209 missing a word: "provides a brief performance evaluation *of* their execution times"
* Line 212 fix capitalization: "7 dof Franka Emika panda" -> "7 DOF Franka Emika Panda"
* Same with like 218: "50 muscle 7 dof" -> "50 muscle, 7 DOF"
* Lines 213-214 missing a word: "computer equipped with *a* 1.90 GHz Intel i7..."
