### 2023-08-21 d1cd861
updated performance analysis
