### 2023-08-21 e0a00b1
updated docs vepoli square problem
