### 2023-08-21 auctus-team/pycapacity#6 @askuric
Hey @sea-bass,

Thanks a lot for your updates, the paper is updated with your comments.
I absolutely agree about the script link, it is not the smartest idea in long-run.

The paper is regenerated in the main issue:
 https://github.com/openjournals/joss-reviews/issues/5670

Let me know if there is anything else you'd like me to do.
