### 2023-08-22 e339e7d
added 1 instead of max for ellipsoids
