### 2023-08-22 4e08735
updated readme with the ellipsoids
