### 2023-08-22 db8cabe
better docs for ellipsoids
