### 2023-08-22 2708f9c
added weighing matrix
