### 2023-08-25 b51041c
some typos and added examples in class definition
