### 2023-08-25 9d7ba02
some typos and added examples in class definition
