### 2023-08-26 auctus-team/pycapacity#9 @askuric
Hey @JHartzer,
Did you maybe have time to look into the changes?
Thanks!
