### 2023-08-27 2e63d3f
changed intersection from - to &, and added examples + referenced reachable space polytpope to pinocchio
