### 2023-08-29 auctus-team/pycapacity#9 @JHartzer
Howdy @askuric, yes I have! Excellent work and thank you for your work. I believe you've addressed all my comments, so I will go ahead and close out this issue and complete my review.
