### 2023-08-30 auctus-team/pycapacity#6 @askuric
Hi @ShravanTata,
Sorry to rush you, but did you maybe have time to look into the made changes?
I'd be happy to hear your feedback.
Thanks!
