### 2023-09-04 7443780
docs prepared for v2.0.1. + color parameter in plot_polytope listed
