### 2023-09-04 91937d1
updated readme with 2.0.1 + link to changelog in the docs
