+++ README.md
@@ -2,22 +2,8 @@
 # An efficient task-space capacity calculation package for robotics and biomechanics
 
 
-> 📢 New version of the pycapacity package is out - version v2.0!
-> 🚨 Beware because there are some breaking changes in the API, please check the changelog for more details.
+> 📢 New version of the pycapacity package is out - version v2.0.1! - [see full changelog](https://auctus-team.github.io/pycapacity/changelog.html)
 
-### changelog v2.0 (05-2023)
-
-
-- Polytope and ellipsoid algorithms now return `Polytope`  and `Ellipsoid`  objects (see [api docs](https://auctus-team.github.io/pycapacity/pycapacity.polytope.html))
-  - `Polytope`  class can be used as a standalone class to find  vertex, halfplane and face representation of a polytope
-- Visualisation tools improved: (see [api docs](https://auctus-team.github.io/pycapacity/pycapacity.visual.html))
-  - New `plot_polytope` function for plotting polytopes
-  - Functions `plot_polytope_vertex` and `plot_polytope_faces` now can receive a polytope object
-  - Function `plot_ellipsoid` now can receive a ellipsoid object
-  - Better management of the `matplotlib` figures and axes, now user can provide an `ax`, `plt` or `figure` to the plotting functions
-- Added support for robot reachable workspace approximation using convex polytopes (see [API docs](https://auctus-team.github.io/pycapacity/pycapacity.robot.html#pycapacity.robot.reachable_space_approximation))
-- Added unit testing + continuous integration
-- Improved docs with sphinx
 
 ## About 
 [![PyPI package](https://img.shields.io/pypi/v/pycapacity)](https://pypi.org/project/pycapacity/) [![Tests](https://github.com/auctus-team/pycapacity/actions/workflows/python-app.yml/badge.svg)](https://github.com/auctus-team/pycapacity/actions/workflows/python-app.yml) ![](https://img.shields.io/pypi/dm/pycapacity?color=blue&label=pip%20downloads)  [![Docs](https://github.com/auctus-team/pycapacity/actions/workflows/main.yml/badge.svg)](https://github.com/auctus-team/pycapacity/actions/workflows/main.yml)
