+++ README.md
@@ -2,7 +2,7 @@
 # An efficient task-space capacity calculation package for robotics and biomechanics
 
 
-> 📢 New version of the pycapacity package is out - version v2.0.1! - [see full changelog](https://auctus-team.github.io/pycapacity/changelog.html)
+> 📢 New version of the pycapacity package is out - version v2.1! - [see full changelog](https://auctus-team.github.io/pycapacity/changelog.html)
 
 
 ## About 
