### 2023-09-04 #5670 @askuric
Hi @danasolav,
I've published the new library version containing all the modifications proposed by the reviewers. The new library version is v2.1.
Please don't hesitate to let me know the next steps. Do you want me to go ahead and make the zenodo submission?
Thanks!
