### 2023-09-11 #5670 @danasolav
@askuric, I have a few minor comments regarding the paper:
- [ ] Please insert the figures after their first reference in the text.
- [ ] Line 101: There's no need to repeat "Yoshikawa". You can keep the citation only.
- [ ] Please ensure all the equations' variables are properly defined in the text. For example, a clear definition of E, J, P, W, M should be given when they are first mentioned in the text (not in the figure caption). The definitions in 168-170 should appear earlier in the text.
- [ ] Line 118: Could you add references to this claim?
- [ ] Please check the missing DOIs and their suggestions in the previous message.
