+++ paper/paper.bib
@@ -27,6 +27,7 @@
 	address      = {Dordrecht},
 	pages        = {475--482},
 	isbn         = {978-90-481-9262-5},
+	doi			 = {10.1007/978-90-481-9262-5_51},
 	editor       = {Lenarcic, Jadran and Stanisic, Michael M.},
 	abstract     = {The available wrench set is the set of wrenches that can be generated at the platform of a parallel manipulator. In a given configuration, this set is known to be a convex polytope and, recently, it has been pointed out that it is in fact a particular type of convex polytope called a zonotope. This paper deals with the case of parallel manipulators having as many or more actuators than degrees of freedom and discusses a characterization of the facets of the available wrench set.}
 }
@@ -39,6 +40,7 @@
 	address      = {Cham},
 	pages        = {45--60},
 	isbn         = {978-3-031-22731-8},
+	doi 		 = {10.1007/978-3-031-22731-8_4},
 	editor       = {Borja, Pablo and Della Santina, Cosimo and Peternel, Luka and Torta, Elena}
 }
 @article{yoshikawa1985manipulability,
@@ -118,7 +120,7 @@
 	volume       = 33,
 	number       = 6,
 	pages        = {829--840},
-	dio          = {10.1007/s10439-005-3320-7}
+	doi          = {10.1007/s10439-005-3320-7}
 }
 @inproceedings{quigley2009ros,
 	title        = {ROS: an open-source Robot Operating System},
@@ -251,9 +253,20 @@
 }
 
 @article{fukuda1997cdd,
-  title={cdd/cdd+ Reference Manual},
-  author={Fukuda, Komei},
-  journal={Institute for Operations Research, ETH-Zentrum},
-  pages={91--111},
-  year={1997}
+	title	=	{cdd/cdd+ Reference Manual},
+	author	=	{Fukuda, Komei},
+	journal	=	{Institute for Operations Research, ETH-Zentrum},
+	pages	=	{91--111},
+	year	=	{1997}
+}
+
+@article{Finotello1998Computation,
+	title={Computation of kinetostatic performances of robot manipulators with polytopes},
+	author={Roberto Finotello and Tiberio Grasso and Giorgio Rossi and Antonio Terribile},
+	journal={Proceedings. 1998 IEEE International Conference on Robotics and Automation (Cat. No.98CH36146)},
+	year={1998},
+	volume={4},
+	pages={3241-3246 vol.4},
+	doi = {10.1109/ROBOT.1998.680928},
+	url={https://ieeexplore.ieee.org/abstract/document/680928/}
 }
