+++ paper/paper.md
@@ -54,17 +54,15 @@ In robotics, task-space physical ability metrics establish the relationship betw
 
 When it comes to characterizing these achievable sets, the two most common approaches are using ellipsoids and polytopes. Ellipsoids are often used to represent the robot's velocity capacity, so called manipulability, while polytopes are mostly used to represent the robot's force capacity. However, both ellipsoids and polytopes can be used to represent any of the task-space physical ability.
 
-![An example manipulability polytope and ellipsoid geometry for a planar $m=2$ robot with $n=2$. The difference between the joint space limits for ellipsoid described with $||\dot{{q}}||_2\leq1$ (orange) and the range limits ${-1}\leq\dot{{q}}\leq{1}$ (blue) is shown on the right. The difference in obtained achievable task-space velocity $\dot{{x}}$ polytope ${P}$ (blue) and ellipsoid ${E}$ (orange) is shown on the right plot. The plots show that both in joint and task-space the ellipsoid metric is an underestimation of the true robot's capacity.](ellip_poly.png){ width=100% }
-
 To compare the ellipsoid and polytope metrics, the example of the manipulability ellipsoid and manipulability polytope can be used. 
 
-The manipulability ellipsoid, proposed by Yoshikawa [@yoshikawa1985manipulability], is defined as the set of all achievable task-space velocities ${\dot{x}}$ for a given robot configuration ${q}$ and joint velocity limits $-1 \leq {\dot{q}}\leq 1$, and it can be expressed as:
+The manipulability ellipsoid $E$, proposed by [@yoshikawa1985manipulability], is defined as the set of all achievable task-space velocities ${\dot{x}}$ for a given robot configuration ${q}$ and joint velocity limits $-1 \leq {\dot{q}}\leq 1$, and it can be expressed as:
 
 \begin{equation}\label{eq:manip_ellipsoid}
 E = \{ \dot{x} ~|~ \dot{x} = J({q})\dot{q}, \quad ||\dot{q}||_2 \leq 1 \}
 \end{equation}
 
-The equivalent polytope representation of the manipulability ellipsoid is the manipulability polytope, which is defined as the set of all achievable task-space velocities ${\dot{x}}$ for a given robot configuration ${q}$ and joint velocity limits $-1 \leq {\dot{q}}\leq 1$, and it can be expressed as:
+The equivalent polytope $P$ representation of the manipulability ellipsoid is the manipulability polytope, which is defined as the set of all achievable task-space velocities ${\dot{x}}$ for a given robot configuration ${q}$ and joint velocity limits $-1 \leq {\dot{q}}\leq 1$, and it can be expressed as:
 
 \begin{equation}\label{eq:manip_polytope}
 P = \{ \dot{x} ~|~ \dot{x} = J({q})\dot{q}, \quad -1 \leq \dot{q} \leq 1 \}
@@ -72,10 +70,12 @@ P = \{ \dot{x} ~|~ \dot{x} = J({q})\dot{q}, \quad -1 \leq \dot{q} \leq 1 \}
 
[…]
