+++ paper/paper.bib
@@ -235,7 +235,7 @@
 
 @article{fukuda2004frequently,
   title={Frequently asked questions in polyhedral computation},
-  author={Fukuda, Komei and others},
+  author={Fukuda, Komei},
   journal={ETH, Zurich, Switzerland},
   volume={85},
   year={2004}
