### 2023-09-11 #5670 @askuric
hey @danasolav ,
Thanks for the comments. I've made a new iteration of the paper by including your remarks.

> - Please insert the figures after their first reference in the text.

Done.

> - Line 101: There's no need to repeat "Yoshikawa". You can keep the citation only.

 Fair enough.

 > - Please ensure all the equations' variables are properly defined in the text. For example, a clear definition of E, J, P, W, M should be given when they are first mentioned in the text (not in the figure caption). The definitions in 168-170 should appear earlier in the text.

I've added the missing definitions and put the definition lines at the end of each subsection which defines the polytopes/ellipsoids.

> - Line 118: Could you add references to this claim?

I've add a reference here in which the authors say explicitly that the polytope computation is more computationally
complex.  Additionally, in the sections that follow the line 118 in our paper, we do explain a bit more in detail the difference in computational complexity and where does it some from.

> - Please check the missing DOIs and their suggestions in the previous message.

The missing DOIs are updated.

Don't hesitate to let me know if you have any other comments.
