### 2023-09-11 ec4e53a
joss info typo
