### 2023-09-11 4066a6c
paper typo in the label of the equations
