### 2023-09-11 #5670 @askuric
Ok, the new version is: v2.1.2
DOI is: 10.5281/zenodo.8333929
Zenodo link is: https://zenodo.org/record/8333929
