### 2023-09-11 #5670 @danasolav
@askuric and co-authors, congratulations! your paper is ready to be accepted and is now in the hands of the Editor-in-Chief.
@sea-bass @ShravanTata @JHartzer, Thank you for your reviews. Your efforts are much appreciated! JOSS is based on volunteers and we couldn't do this without you.
