### 2023-09-12 #5670 @Kevin-Mattheus-Moerman
@askuric I am the AEiC on this track and here to help process the final steps. I have checked this review, your software repository, the archive link, and the paper, and most seems in order. Below is one point which requires your attention:

- [ ] Although the above boxes for contributing guidelines are ticked, I cannot see this in your repository. Please can you add contributing guidelines to your software repository? You may choose for instance to have a simple section in the README detailing how folks can contribute/post issues, or you could have a dedicated CONTRIBUTING.md document and link to it in the README. This website contains some examples of the latter: https://contributing.md/example/
