### 2023-09-12 #5670 @sea-bass
Does it count if the contributing guidelines are in the documentation, but not as a specific markdown file in the repo?
https://auctus-team.github.io/pycapacity/contirbuting.html
