### 2023-09-12 #5670 @Kevin-Mattheus-Moerman
@sea-bass, yes that is fine actually. In that case, I can proceed here with acceptance. However, I would recommend the following:

- [ ] Link to these contributing guidelines in your README
- [ ] You may want to fix the typo in the URL which currently says `contirbuting`
