### 2023-09-12 #5670 @askuric
Awesome!
I've gone and added a short contributing.md as well as the short section in the README.

Thanks a lot guys @sea-bass and @Kevin-Mattheus-Moerman!
