### 2023-09-12 4b4992c
readme shrink and added citation
