### 2023-09-12 #5670 @Kevin-Mattheus-Moerman
@openjournals/dev this DOI doesn't resolve yet. Do I just repeat the accept step?
